Moved the IUserProvider functionality into IRequestInformationProvider because this is all about extracting information from the request
