+++ DDDEastAnglia.Tests/Voting/VotingTestBase.cs
@@ -18,7 +18,6 @@ namespace DDDEastAnglia.Tests.Voting
         protected VoteController Controller;
         protected ITimeProvider TimeProvider;
         protected IRequestInformationProvider RequestInformationProvider;
-        protected IUserProvider UserProvider;
         protected DateTime SimulatedNow;
 
         [SetUp]
@@ -43,15 +42,8 @@ namespace DDDEastAnglia.Tests.Voting
             RequestInformationProvider = Substitute.For<IRequestInformationProvider>();
             SetRequestInformationProviderExpectations(RequestInformationProvider);
 
-            UserProvider = Substitute.For<IUserProvider>();
-            SetUserProviderExpectations(UserProvider);
 
-            Controller = new VoteController(CookieRepository, VoteRepository, SessionRepository, EventRepository, TimeProvider, RequestInformationProvider, UserProvider);
-        }
-
-        protected virtual void SetUserProviderExpectations(IUserProvider userProvider)
-        {
-        
+            Controller = new VoteController(CookieRepository, VoteRepository, SessionRepository, EventRepository, TimeProvider, RequestInformationProvider);
         }
 
         protected virtual void SetCookieRepositoryExpectations(IVotingCookieRepository repository)
