+++ DDDEastAnglia/DDDEastAnglia.csproj
@@ -239,9 +239,7 @@
     <Compile Include="DataModel\Vote.cs" />
     <Compile Include="GlimpseSecurityPolicy.cs" />
     <Compile Include="Helpers\Context\HttpContextRequestInformationProvider.cs" />
-    <Compile Include="Helpers\Context\HttpContextUserProvider.cs" />
     <Compile Include="Helpers\IRequestInformationProvider.cs" />
-    <Compile Include="Helpers\IUserProvider.cs" />
     <Compile Include="Helpers\TimeProvider.cs" />
     <Compile Include="HomeController.generated.cs">
       <DependentUpon>T4MVC.tt</DependentUpon>
