+++ DDDEastAnglia/Helpers/Context/HttpContextRequestInformationProvider.cs
@@ -1,12 +1,17 @@
+using System.Linq;
 using System.Net;
 using System.Text.RegularExpressions;
 using System.Web;
+using DDDEastAnglia.DataAccess;
+using DDDEastAnglia.Models;
 
 namespace DDDEastAnglia.Helpers.Context
 {
     public class HttpContextRequestInformationProvider : IRequestInformationProvider
     {
+        private DDDEAContext context = new DDDEAContext();
         private static readonly Regex IPV4AddressMatch = new Regex(@"\b(?<IPAddress>((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?))\b", RegexOptions.Compiled);
+
         public string GetIPAddress()
         {
             var request = HttpContext.Current.Request;
@@ -14,6 +19,16 @@ namespace DDDEastAnglia.Helpers.Context
                    ?? MatchIPAddress(request.Headers["HTTP_VIA"])
                    ?? MatchIPAddress(request.Headers["HTTP_PROXY_CONNECTION"])
                    ?? MatchIPAddress(request.UserHostAddress);
+        }
+
+        public bool IsLoggedIn()
+        {
+            return !string.IsNullOrWhiteSpace(HttpContext.Current.User.Identity.Name);
+        }
+
+        public UserProfile GetCurrentUser()
+        {
+            return context.UserProfiles.FirstOrDefault(profile => profile.UserName == HttpContext.Current.User.Identity.Name);
         }
 
         public static string MatchIPAddress(string potentialIPAddress)
