+++ DDDEastAnglia/Helpers/IRequestInformationProvider.cs
@@ -1,7 +1,11 @@
-﻿namespace DDDEastAnglia.Helpers
+﻿using DDDEastAnglia.Models;
+
+namespace DDDEastAnglia.Helpers
 {
     public interface IRequestInformationProvider
     {
         string GetIPAddress();
+        bool IsLoggedIn();
+        UserProfile GetCurrentUser();
     }
 }
