--- DDDEastAnglia/Helpers/IUserProvider.cs
@@ -1,10 +1,0 @@
-﻿using DDDEastAnglia.Models;
-
-namespace DDDEastAnglia.Helpers
-{
-    public interface IUserProvider
-    {
-        bool IsLoggedIn();
-        UserProfile GetCurrentUser();
-    }
-}
